mdnsd: Remove 'net_raw' GID.
The 'net_raw' GID is an Android addition that does not exist upstream.
It grants the CAP_NET_RAW capability. However, it does not appear to be
required for mdnsd. There are no raw sockets being created in the code:
mdnsresponder$ grep -R "socket(" * shows no use of raw sockets.

As per 'man capabilities(7)', CAP_NET_RAW is used for:
"""
* use RAW and PACKET sockets;
* bind to any address for transparent proxying.
"""

Removing this use of 'net_raw' will allow us to drop the kernel patches
that grant CAP_NET_RAW with the 'net_raw' GID.

Bug: 33938230
Test: 'start mdnsd' works.
Test: Install "Bonjour Browser" app, do a scan.
Test: 'adb logcat -b all -d | grep mDNS' shows mDNSResponder starting.
Change-Id: Idb67788d56115c1b36a97c1962fd03eb26e82a68