Add implicit output instead of double rules
This used to define two rules to build EventLog.h / EventLog.o, which
would sometimes run at the same time, corrupting each others results.

The intended behavior was to define a single rule that built both output
files, but it's impossible to represent this in normal make. I've added
an extension to Kati to support this, since our underlying build engine
(Ninja) already supports this.

Bug: 63693187
Test: Inspect build.ninja file
Test: mmma external/mdnsresponder
Change-Id: I08d31be1f391f0cc47e27150c5d382bfb91aad2e