Compile the mdns codebase with -fwrapv.
The timekeeping code in mdnsd relies on signed integer overflow
wrapping, otherwise it behaves incorrectly (specifically, it gets
stuck in an infinite loop when its timer ticks reach 0x7fffffff).

For safety, instead of fixing this one case, assume that there
are other parts of the mdns codebase that rely on signed integer
overflow wrapping, and compile with the appropriate compiler
option.

Also suppress a noisy compiler warning.

Bug: 65051991
Bug: 63335997

Test: Manual
Change-Id: I7a737ea61114cb080717a8ef1b269cc3a48da4d5