Fix UWB start/stop ranging causing mDNS drop
Current mDNS implementation closes and re-opens all of the network
interfaces when any of the existing interfaces changes. Since UWB on
start/stop ranging modifies its interface, mDNS detects that as a
significant enough change to restart all of its interfaces. This change
prevents that behaviour since the UWB interface is not related to anything
what mDNS does. If the detected interface change is neither an
IPv4 or IPv6 capable interface and if the interface type is
ARPHRD_IEEE802154 which is the IEEE 802.15.4 PHY and MAC standard that
UWB uses, then mDNS will ignore it.

Bug:265207453
Test: Verified manually after this fix is applied interfaces no longer
restart on UWB start/stop ranging.

Change-Id: I1c070d989769752fda92bbe9308ca5bfc7c2d3e5